May19/requirements.txt
google-generativeai
group_chat
gemini_utils
dotenv